# eims-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExamFlow EIMS API
  description: |
    Exam Information Management System API for UP Tacloban.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requests per 15 minutes per IP
    - Rate limit headers included in responses
    
    ## WebSocket
    Real-time updates available at `ws://localhost:3001/ws?token=<access_token>`
  version: 1.0.0
  contact:
    name: ExamFlow Support
    email: support@examflow.up.edu.ph
  license:
    name: Educational Use
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://examflow.up.edu.ph/api/v1
    description: Production server
...
paths:
...
  /exams:
    get:
      tags: [Exams]
      summary: List exams
      description: Get paginated list of exams (filtered by role)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ExamStatus'
      responses:
        '200':
          description: List of exams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamListResponse'
    post:
      tags: [Exams]
      summary: Create exam
      description: Create a new exam (Faculty/Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExamRequest'
      responses:
        '201':
          description: Exam created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /exams/{id}:
    get:
      tags: [Exams]
      summary: Get exam details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Exam details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Exams]
      summary: Update exam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExamRequest'
      responses:
        '200':
          description: Exam updated
    delete:
      tags: [Exams]
      summary: Delete exam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Exam deleted
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread count
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unread count
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    ExamType:
      type: string
      enum: [WRITTEN, ORAL, PRACTICAL, ONLINE, TAKE_HOME]
    ExamStatus:
      type: string
      enum: [DRAFT, SCHEDULED, ONGOING, COMPLETED, CANCELLED]
...
    Exam:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        courseId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ExamType'
        status:
          $ref: '#/components/schemas/ExamStatus'
        totalPoints:
          type: integer
        passingScore:
          type: integer
        guidelines:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateExamRequest:
      type: object
      required: [title, courseId, type, totalPoints, passingScore]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
        courseId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ExamType'
        totalPoints:
          type: integer
          minimum: 1
        passingScore:
          type: integer
          minimum: 0
        guidelines:
          type: string
    UpdateExamRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/ExamType'
        status:
          $ref: '#/components/schemas/ExamStatus'
        totalPoints:
          type: integer
        passingScore:
          type: integer
        guidelines:
          type: string
...
    ExamResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Exam'
    ExamListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Exam'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'